1. remove dot from float .

+++ hasoffer.job/src/main/java/hasoffer/job/bean/deal/CheckGetPriceOffDealJobBean.java
@@ -20,6 +20,7 @@ import org.springframework.scheduling.quartz.QuartzJobBean;
 
 import javax.annotation.Resource;
 import java.io.File;
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
@@ -149,11 +150,11 @@ public class CheckGetPriceOffDealJobBean extends QuartzJobBean {
                     if (newPrice >= minPrice) {//100%-110%
                         //如果现价不低于史低价
                         //文案 Rs.现价 is almost history lowest price(History lowest price is Rs.更新前的史低价). Click here to check price history（点击此行展示价格走势浮层）. Good offer always expire in hours.Good time to get it,Hurry up!
-                        sb.append("Rs.").append(newPrice).append(" is almost history lowest price(History lowest price is Rs.").append(minPrice).append(").Click here to check price history.Good offer always expire in hours.Good time to get it,Hurry up!");
+                        sb.append("Rs.").append(BigDecimal.valueOf(newPrice)).append(" is almost history lowest price(History lowest price is Rs.").append(minPrice).append(").Click here to check price history.Good offer always expire in hours.Good time to get it,Hurry up!");
                     } else {//小于100%
                         //否则
                         //Rs.现价 is the newest history lowest price(Previous lowest price is Rs.更新前的史低价).Click here to check price history（以高亮可点击文案展示点击唤出价格走势浮层 具体逻辑见price history）. Good offer always expire in hours.Good time to get it,Hurry up!
-                        sb.append("Rs.").append(newPrice).append(" is newest history lowest price(Previous lowest price is Rs.").append(minPrice).append(").Click here to check price history.Good offer always expire in hours.Good time to get it,Hurry up!");
+                        sb.append("Rs.").append(BigDecimal.valueOf(newPrice)).append(" is newest history lowest price(Previous lowest price is Rs.").append(minPrice).append(").Click here to check price history.Good offer always expire in hours.Good time to get it,Hurry up!");
                     }
 
                     appdeal.setDescription(sb.toString());
